+++ VideoApp/VideoApp/NewUI/Views/Core/DisplayNameFactory.swift
