+++ VideoApp/VideoApp/NewUI/Views/Core/ErrorAlert.swift
