+++ VideoApp/VideoApp/NewUI/Views/GridLayout/SpeakerGridView.swift
