+++ VideoApp/VideoApp/NewUI/Views/RoomStatus/RoomStatusView.swift
@@ -4,7 +4,7 @@
 
 import SwiftUI
 
-struct StreamStatusView: View {
+struct RoomStatusView: View {
     let streamName: String
     @Binding var streamState: StreamManager.State
     
@@ -20,17 +20,18 @@ struct StreamStatusView: View {
     }
 }
 
-struct StreamStatusView_Previews: PreviewProvider {
+struct RoomStatusView_Previews: PreviewProvider {
     static var previews: some View {
         Group {
-            StreamStatusView(streamName: "Short room name", streamState: .constant(.connected))
+            RoomStatusView(streamName: "Short room name", streamState: .constant(.connected))
                 .previewDisplayName("Short Room Name")
-            StreamStatusView(
+            RoomStatusView(
                 streamName: "A very long room name that does not fit and is truncated",
                 streamState: .constant(.connected)
             )
                 .previewDisplayName("Long Room Name")
         }
+        .frame(width: 400)
         .previewLayout(.sizeThatFits)
     }
 }
