+++ VideoApp/VideoApp/ViewControllers/Lobby/LobbyViewController.swift
@@ -132,7 +132,7 @@ class LobbyViewController: UIViewController {
 //        performSegue(withIdentifier: "roomSegue", sender: self)
         
         let config = StreamConfig(streamName: roomName, userIdentity: AuthStore.shared.userDisplayName)
-        let controller = UIHostingController(rootView: StreamView(config: config))
+        let controller = UIHostingController(rootView: RoomScreenView(config: config))
         present(controller, animated: true)
     }
